% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Etna_boundary.r
\docType{data}
\name{Etna_boundary}
\alias{Etna_boundary}
\title{Dataset: bounding polygon representing a study area on Mount Etna (Sicily, Italy)}
\format{
SpatialPolygonDataFrame
}
\usage{
data(Etna_boundary)
}
\description{
A SpatialPolygonDataFrame representing an area on Mount Etna (Sicily, Italy), to be used to download elevation data.
}
\keyword{datasets}
