% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Etna_end_location.r
\docType{data}
\name{Etna_end_location}
\alias{Etna_end_location}
\title{Dataset: locations on Mount Etna (Sicily, Italy)}
\format{
SpatialPointsDataFrame
}
\usage{
data(Etna_end_location)
}
\description{
A SpatialPointsDataFrame representing spots on Mount Etna (Sicily, Italy), to be used as destination locations for least-cost paths calculation.
}
\keyword{datasets}
