% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Etna_start_location.r
\docType{data}
\name{Etna_start_location}
\alias{Etna_start_location}
\title{Dataset: location on Mount Etna (Sicily, Italy)}
\format{
SpatialPointsDataFrame
}
\usage{
data(Etna_start_location)
}
\description{
A SpatialPointsDataFrame representing a spot on Mount Etna, to be used as start location for least-cost paths calculation.
}
\keyword{datasets}
