% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malta_dtm_40.r
\docType{data}
\name{malta_dtm_40}
\alias{malta_dtm_40}
\title{Dataset: Malta DTM (40m cell size)}
\format{
RasterLayer
}
\usage{
data(malta_dtm_40)
}
\description{
A RasterLayer representing a Digital Terrain Model of Malta (40m resolution).\cr
}
\keyword{datasets}
