% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/springs.r
\docType{data}
\name{springs}
\alias{springs}
\title{Dataset: location of springs in Malta}
\format{
SpatialPointsDataFrame
}
\usage{
data(springs)
}
\description{
A SpatialPointsDataFrame representing the location of spring-related toponyms in Malta.
}
\keyword{datasets}
