% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volc.r
\docType{data}
\name{volc}
\alias{volc}
\title{Dataset: raster dataset representing the elevation of the volcano Maunga Whau (Auckland, New Zealand)}
\format{
RasterLayer
}
\usage{
data(volc)
}
\description{
A RasterLayer representing a DTM of the volcano Maunga Whau (Auckland, New Zealand).
}
\keyword{datasets}
